delete

+++ components/message/store.js
@@ -23,19 +23,25 @@ export default class Store {
     }
 
     async getById(id) {
-        return await Model.findOne({_id: id})
+        if (await this.ifExist(id)) return await Model.findOne({ _id: id })
     }
 
     async update(id, message, updatedTime) {
-        return await Model.findOneAndUpdate(
+        if (await this.ifExist(id)) return await Model.findOneAndUpdate(
             { _id: id },
             { $set: { message, date: updatedTime } },
             { new: true }
         )
     }
     async delete(id) {
-        return await Model.findByIdAndDelete(
-            { _id: id }
-        )
+        if (await this.ifExist(id)) return await Model.findByIdAndDelete({ _id: id })
+    }
+    
+    async ifExist(id) {
+        const validateId = await Model.exists({ _id: id })
+        if (!validateId) {
+            throw new Error('Id no existe')
+        }
+        return true
     }
 }
